Work in progress

+++ Assets/CompanyName/RamRetribution/Scripts/Interfaces/IArmor.cs
@@ -4,6 +4,6 @@ namespace CompanyName.RamRetribution.Scripts.Interfaces
 {
     public interface IArmor
     {
-        public float ReduceDamage(AttackType type ,float damage);
+        public int ReduceDamage(AttackType type ,float damage);
     }
 }
